% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_preprocess.R
\name{calculate_air_pollution}
\alias{calculate_air_pollution}
\title{Calculate air pollution exposure averages}
\usage{
calculate_air_pollution(df, pollutants = c("NO2", "PM10", "PM2.5", "NOx"))
}
\arguments{
\item{df}{A data.table containing air pollution columns}

\item{pollutants}{Character vector of pollutants to calculate.
Available: "NO2", "PM10", "PM2.5", "NOx"}
}
\value{
A data.table with averaged pollution columns
}
\description{
Computes averaged air pollution exposures from multiple time points.
}
