% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolomics.R
\name{classify_metabolites}
\alias{classify_metabolites}
\title{Classify UK Biobank metabolite names}
\usage{
classify_metabolites(metabolites)
}
\arguments{
\item{metabolites}{Character vector of metabolite names.}
}
\value{
A data.frame with \code{metabolite}, \code{category}, and \code{metaboanalyst_name}.
}
\description{
Classify metabolite-like names into broad groups used by the metabolomics
ORA workflow. Small molecules can be mapped to MetaboAnalyst-compatible names,
whereas lipoprotein subclass measures, lipid aggregate measures, and proteins
are retained in the mapping table but are not passed to small-molecule ORA by
default.
}
\examples{
classify_metabolites(c("Alanine", "LDL Cholesterol", "Apolipoprotein B"))

}
