% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{coef.mediation_result}
\alias{coef.mediation_result}
\title{Extract Coefficients from Mediation Results}
\usage{
\method{coef}{mediation_result}(object, ...)
}
\arguments{
\item{object}{An object of class "mediation_result".}

\item{...}{Additional arguments (unused).}
}
\value{
A data.frame with effect estimates.
}
\description{
Extract effect estimates from mediation analysis results.
}
