% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtableone.R
\name{create_baseline_table}
\alias{create_baseline_table}
\title{Create a baseline table comparing cases and controls under different conditions.}
\usage{
create_baseline_table(
  data,
  case_col,
  factor_cols = NULL,
  continuous_cols = NULL,
  test = FALSE
)
}
\arguments{
\item{data}{a data.table containing the baseline characteristics and case/control status.}

\item{case_col}{the name of the column indicating case/control status (1 for cases, 0 for controls).}

\item{factor_cols}{a vector of column names that are factors (categorical variables)}

\item{continuous_cols}{a vector of column names that are continuous variables.}

\item{test}{whether to perform statistical tests comparing cases and controls for each variable (default: FALSE).}
}
\value{
a list containing table one information.
}
\description{
Create a baseline table comparing cases and controls under different conditions.
}
\references{
https://github.com/kaz-yos/tableone
}
