% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{.create_shap_predict_wrapper}
\alias{.create_shap_predict_wrapper}
\title{Create prediction wrapper for SHAP}
\usage{
.create_shap_predict_wrapper(object)
}
\description{
Create prediction wrapper for SHAP
}
\keyword{internal}
