% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{.fit_logistic}
\alias{.fit_logistic}
\title{Fit Logistic/Linear Regression}
\usage{
.fit_logistic(formula, data, task, params, verbose, ...)
}
\description{
Fit Logistic/Linear Regression
}
\keyword{internal}
