% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_preprocess.R
\name{.get_variable_mapping}
\alias{.get_variable_mapping}
\title{Get default variable to UKB field ID mapping}
\usage{
.get_variable_mapping()
}
\value{
A named list with variable mappings
}
\description{
Get default variable to UKB field ID mapping
}
\keyword{internal}
