% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parse formula to get response and predictors}
\usage{
.parse_formula(formula, data)
}
\description{
Parse formula to get response and predictors
}
\keyword{internal}
