% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{filter_death_codes}
\alias{filter_death_codes}
\title{Filter Death Records by ICD-10 Code Pattern}
\usage{
filter_death_codes(death_long, pattern, disease_label)
}
\arguments{
\item{death_long}{A data.table from \code{\link{parse_death_records}}.}

\item{pattern}{Regular expression pattern for ICD-10 death codes.}

\item{disease_label}{Disease name label to assign to matched records.}
}
\value{
A data.table with filtered records and added \code{disease} column.
}
\description{
Filters death cause records using regular expression pattern matching.
}
\keyword{internal}
