% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcs4.R
\name{filter_opcs4_codes}
\alias{filter_opcs4_codes}
\title{Filter OPCS4 Procedure Records by Code Pattern}
\usage{
filter_opcs4_codes(opcs4_long, pattern, disease_label)
}
\arguments{
\item{opcs4_long}{A data.table from \code{\link{parse_opcs4_procedures}}.}

\item{pattern}{Regular expression pattern for OPCS4 procedure codes.}

\item{disease_label}{Disease name label to assign to matched records.}
}
\value{
A data.table with filtered records and added \code{disease} column.
}
\description{
Filters OPCS4 procedure records using regular expression pattern matching.
}
\keyword{internal}
