% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_workflow.R
\name{plot.ukb_ml_flow}
\alias{plot.ukb_ml_flow}
\title{Plot a UKB ML Flow Object}
\usage{
\method{plot}{ukb_ml_flow}(x, type = c("roc", "shap_beeswarm"), ...)
}
\arguments{
\item{x}{A \code{ukb_ml_flow} object.}

\item{type}{Plot type: \code{"roc"} or \code{"shap_beeswarm"}.}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\value{
A ggplot2 object.
}
\description{
Plot a UKB ML Flow Object
}
