% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_workflow.R
\name{plot.ukb_ml_flow_compare}
\alias{plot.ukb_ml_flow_compare}
\title{Plot a UKB ML Flow Comparison Object}
\usage{
\method{plot}{ukb_ml_flow_compare}(x, type = c("roc"), ...)
}
\arguments{
\item{x}{A \code{ukb_ml_flow_compare} object.}

\item{type}{Plot type. Currently \code{"roc"}.}

\item{...}{Additional arguments passed to \code{\link{plot_ml_roc_compare}}.}
}
\value{
A ggplot2 object.
}
\description{
Plot a UKB ML Flow Comparison Object
}
