% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabolomics.R
\name{plot_metabolite_ora_barplot}
\alias{plot_metabolite_ora_barplot}
\title{Plot metabolite ORA results as a bar plot}
\usage{
plot_metabolite_ora_barplot(
  x,
  top_n = 15,
  p_col = "pvalue",
  pathway_col = "pathway",
  fill_color = "#2F6FA3"
)
}
\arguments{
\item{x}{A data.frame returned by \code{run_metabolite_ora()$ora_result} or a
\code{ukb_metabolite_ora} object.}

\item{top_n}{Number of pathways to show. Default \code{15}.}

\item{p_col}{P-value column used for ordering and color. Default \code{"pvalue"}.}

\item{pathway_col}{Column containing pathway names. Default \code{"pathway"}.}

\item{fill_color}{Bar color.}
}
\value{
A ggplot object.
}
\description{
Plot metabolite ORA results as a bar plot
}
