% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_calibration}
\alias{plot_ml_calibration}
\title{Plot Calibration Curve}
\usage{
plot_ml_calibration(object, title = "Calibration Curve", ...)
}
\arguments{
\item{object}{A ukb_ml_calibration object from ukb_ml_calibration()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create calibration curve plot showing predicted vs observed probabilities.
}
