% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_dca}
\alias{plot_ml_dca}
\title{Plot Decision Curve Analysis}
\usage{
plot_ml_dca(object, title = "Decision Curve Analysis", ...)
}
\arguments{
\item{object}{A ukb_ml_dca object from ukb_ml_dca()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a Decision Curve Analysis plot showing net benefit of the model
compared to treat-all and treat-none strategies.
}
