% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_gain}
\alias{plot_ml_gain}
\title{Plot Gain Curve}
\usage{
plot_ml_gain(object, title = "Gain Curve", ...)
}
\arguments{
\item{object}{A ukb_ml_gain_lift object from ukb_ml_gain_lift()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a Gain curve plot comparing model targeting against random selection.
}
