% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_lift}
\alias{plot_ml_lift}
\title{Plot Lift Curve}
\usage{
plot_ml_lift(object, title = "Lift Curve", ...)
}
\arguments{
\item{object}{A ukb_ml_gain_lift object from ukb_ml_gain_lift()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a Lift curve plot showing the ratio of model vs random targeting.
}
