% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_pr}
\alias{plot_ml_pr}
\title{Plot PR Curve}
\usage{
plot_ml_pr(object, title = "PR Curve", ...)
}
\arguments{
\item{object}{A ukb_ml_pr object from ukb_ml_pr()}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a Precision-Recall curve plot with AUPRC annotation.
}
