% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ml_roc}
\alias{plot_ml_roc}
\title{Plot ROC Curves}
\usage{
plot_ml_roc(object, ci_alpha = 0.2, title = "ROC Curve", ...)
}
\arguments{
\item{object}{A ukb_ml_roc object from ukb_ml_roc()}

\item{ci_alpha}{Alpha for confidence interval ribbon (default 0.2)}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create ROC curve plot for one or more ML models.
}
