% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_shap_force}
\alias{plot_shap_force}
\title{Plot SHAP Force (Waterfall)}
\usage{
plot_shap_force(object, row_id = 1, max_features = 10, title = NULL, ...)
}
\arguments{
\item{object}{A ukb_shap object}

\item{row_id}{Row index to explain (default 1)}

\item{max_features}{Maximum features to show (default 10)}

\item{title}{Plot title}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a waterfall plot showing feature contributions for a single prediction.
}
