% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{print.mediation_result}
\alias{print.mediation_result}
\title{Print Method for Mediation Results}
\usage{
\method{print}{mediation_result}(x, ...)
}
\arguments{
\item{x}{An object of class "mediation_result".}

\item{...}{Additional arguments passed to summary.}
}
\value{
Invisibly returns \code{x}, the original mediation result object.
}
\description{
Print mediation analysis results.
}
