% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{run_correlation}
\alias{run_correlation}
\title{Calculate correlation between variables}
\usage{
run_correlation(df, vars, method = "pearson", threshold = 0.7)
}
\arguments{
\item{df}{A data.frame or data.table containing the variables of interest.}

\item{vars}{A character vector of column names for which to calculate the correlation matrix.}

\item{method}{The method to use for calculating correlation. Options are "pearson", "spearman", or "kendall". Default is "pearson".}

\item{threshold}{Numeric value between 0 and 1. If specified, the variables with absolute correlation above this threshold will be highlighted in the output. Default is 0.7.}
}
\value{
A correlation matrix of the specified variables.
}
\description{
Before the formal regression analysis, it can be useful to check the correlation between variables.
This function calculates the correlation matrix for a set of specified variables, which can help identify potential multicollinearity issues or inform variable selection.
}
