% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{ukb_ml_compare}
\alias{ukb_ml_compare}
\title{Compare Multiple ML Models}
\usage{
ukb_ml_compare(
  ...,
  models = list(),
  metrics = NULL,
  test_data = NULL,
  plot = TRUE
)
}
\arguments{
\item{...}{ukb_ml objects to compare}

\item{models}{Alternative: list of ukb_ml objects}

\item{metrics}{Metrics to compare}

\item{test_data}{Optional common test data}

\item{plot}{Whether to create comparison plot}
}
\value{
ukb_ml_compare object with comparison results
}
\description{
Compare performance of multiple trained ML models.
}
