% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluate.R
\name{ukb_ml_confusion}
\alias{ukb_ml_confusion}
\title{Confusion Matrix}
\usage{
ukb_ml_confusion(object, newdata = NULL, threshold = 0.5, plot = TRUE, ...)
}
\arguments{
\item{object}{A ukb_ml object}

\item{newdata}{Optional new data}

\item{threshold}{Classification threshold (default 0.5)}

\item{plot}{Whether to create confusion matrix plot}

\item{...}{Additional arguments}
}
\value{
ukb_ml_confusion object
}
\description{
Generate confusion matrix for classification model.
}
