% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluate.R
\name{ukb_ml_gain_lift}
\alias{ukb_ml_gain_lift}
\title{Gain and Lift Curve Analysis}
\usage{
ukb_ml_gain_lift(object, newdata = NULL, plot = TRUE, n_bins = 10, ...)
}
\arguments{
\item{object}{A ukb_ml object}

\item{newdata}{Optional new data}

\item{plot}{Whether to create gain and lift plots (default TRUE)}

\item{n_bins}{Number of bins / deciles (default 10)}

\item{...}{Additional arguments}
}
\value{
A ukb_ml_gain_lift object with field data containing:
decile, population_pct, positive_capture_pct, gain, lift
}
\description{
Compute Gain and Lift curves for a binary classification model by
ranking predictions into decile bins.
}
