% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{ukb_ml_importance}
\alias{ukb_ml_importance}
\title{Get Variable Importance}
\usage{
ukb_ml_importance(object, type = NULL, ...)
}
\arguments{
\item{object}{A ukb_ml object}

\item{type}{Importance type (model-specific)}

\item{...}{Additional arguments}
}
\value{
Data frame with variable importance scores
}
\description{
Extract variable importance from a trained ML model.
}
