% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{ukb_ml_predict}
\alias{ukb_ml_predict}
\title{Predict from ML Model}
\usage{
ukb_ml_predict(
  object,
  newdata = NULL,
  type = c("response", "prob", "class", "link"),
  ...
)
}
\arguments{
\item{object}{A ukb_ml object from ukb_ml_model()}

\item{newdata}{Optional new data for prediction. If NULL, uses test data.}

\item{type}{Prediction type: "response", "prob", "class", "link"}

\item{...}{Additional arguments}
}
\value{
Predictions as vector or matrix
}
\description{
Generate predictions from a trained ukb_ml model.
}
