% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_survival.R
\name{ukb_ml_survival_importance}
\alias{ukb_ml_survival_importance}
\title{Get Variable Importance for Survival Model}
\usage{
ukb_ml_survival_importance(object, ...)
}
\arguments{
\item{object}{A ukb_ml_surv object}

\item{...}{Additional arguments}
}
\value{
Data frame with variable importance
}
\description{
Get Variable Importance for Survival Model
}
