% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_workflow.R
\name{ukb_protein_annotation}
\alias{ukb_protein_annotation}
\title{Annotate Olink-style protein variables}
\usage{
ukb_protein_annotation(
  variables,
  protein_prefix = "^olink_instance_0[.]",
  drop_unmapped = FALSE
)
}
\arguments{
\item{variables}{Protein variable names.}

\item{protein_prefix}{Regular expression prefix removed from variables.}

\item{drop_unmapped}{Passed to \code{\link[=protein_to_gene_symbol]{protein_to_gene_symbol()}}.}
}
\value{
A data.frame with variable, protein_clean, gene_symbol, and
mapping_source.
}
\description{
Annotate Olink-style protein variables
}
