% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{ukb_shap_force}
\alias{ukb_shap_force}
\title{SHAP Force Plot Data}
\usage{
ukb_shap_force(object, row_id = 1, max_features = 10, ...)
}
\arguments{
\item{object}{A ukb_shap object}

\item{row_id}{Row index to explain}

\item{max_features}{Maximum features to show}

\item{...}{Additional arguments}
}
\value{
Data frame with feature contributions for the observation
}
\description{
Get SHAP contribution data for a single observation (force plot).
}
