% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_pool.R
\name{create_imputation_list}
\alias{create_imputation_list}
\title{Create an imputationList Object}
\usage{
create_imputation_list(datasets, validate = TRUE)
}
\arguments{
\item{datasets}{A list of data.frames (imputed datasets).}

\item{validate}{Logical; whether to validate that all datasets have the
same structure. Default TRUE.}
}
\value{
An \code{imputationList} object.
}
\description{
Converts a list of data.frames to a \code{imputationList} object for use
with mitools functions.
}
\examples{
\dontrun{
mi_list <- create_imputation_list(list(data1, data2, data3))
models <- with(mi_list, lm(y ~ x))
}

}
