% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_definitions.R
\name{get_predefined_diseases}
\alias{get_predefined_diseases}
\title{Get Predefined Disease Definitions}
\usage{
get_predefined_diseases()
}
\value{
A named list of disease definition objects.
}
\description{
Returns a list of commonly used cardiovascular and metabolic disease
definitions with validated ICD-10, ICD-9, and self-report code mappings.
}
\details{
Included diseases:
\describe{
\item{AA}{Aortic Aneurysm (I71, 441)}
\item{TAA}{Thoracic Aortic Aneurysm}
\item{AAA}{Abdominal Aortic Aneurysm}
\item{CVD}{Cardiovascular Disease}
\item{MI}{Myocardial Infarction}
\item{HF}{Heart Failure}
\item{Stroke}{Stroke (ischemic and hemorrhagic)}
\item{Hypertension}{Essential and secondary hypertension}
\item{Diabetes}{Diabetes Mellitus (all types)}
\item{Vascular_Disease}{Peripheral vascular disease}
}
}
\examples{
\dontrun{
diseases <- get_predefined_diseases()
my_diseases <- diseases[c("AA", "Hypertension", "Diabetes")]
}

}
