% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{plot_enrichment_lollipop}
\alias{plot_enrichment_lollipop}
\title{Plot enrichment results as a lollipop chart via TCMDATA}
\usage{
plot_enrichment_lollipop(x, ...)
}
\arguments{
\item{x}{An \code{enrichResult} object, or a list containing \code{ora_result}.}

\item{...}{Additional arguments passed to \code{TCMDATA::gglollipop()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
A thin wrapper around \code{TCMDATA::gglollipop()} for enrichment results. This
function accepts either a raw \code{enrichResult} object or a list returned by one
of the proteomics ORA helpers in this package.
}
\examples{
\dontrun{
go_res <- run_protein_ora(
  c("olink_instance_0.il6", "olink_instance_0.tnf", "olink_instance_0.crp"),
  ont = "ALL"
)
plot_enrichment_lollipop(go_res)
}

}
