% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_survival.R
\name{ukb_ml_survival_predict}
\alias{ukb_ml_survival_predict}
\title{Predict from Survival ML Model}
\usage{
ukb_ml_survival_predict(
  object,
  newdata = NULL,
  times = c(1, 3, 5, 10),
  type = c("survival", "risk", "chf"),
  ...
)
}
\arguments{
\item{object}{A ukb_ml_surv object}

\item{newdata}{Optional new data}

\item{times}{Time points for survival prediction}

\item{type}{Prediction type: "risk", "survival", "chf" (cumulative hazard)}

\item{...}{Additional arguments}
}
\value{
Matrix of predictions (observations x time points)
}
\description{
Generate predictions from a survival ML model.
}
