% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_survival.R
\name{ukb_ml_survival_shap}
\alias{ukb_ml_survival_shap}
\title{SHAP Values for Survival Models}
\usage{
ukb_ml_survival_shap(
  object,
  data = NULL,
  time_point = 5,
  nsim = 50,
  sample_n = NULL,
  seed = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A ukb_ml_surv object}

\item{data}{Data for SHAP computation}

\item{time_point}{Time point for SHAP calculation}

\item{nsim}{Number of Monte Carlo samples}

\item{sample_n}{Subsample size}

\item{seed}{Random seed}

\item{verbose}{Print progress}

\item{...}{Additional arguments}
}
\value{
A ukb_shap object
}
\description{
Compute SHAP values for survival ML models at a specific time point.
}
